fix: rename ProjectSelect and selection bug
